+++ private/UIKitCore/UIKBBackdropView.h
@@ -16,7 +16,7 @@
 @property(retain, nonatomic) UIImage *imageForCorners; // @synthesize imageForCorners=_imageForCorners;
 - (void)_setRenderConfig:(id)arg1;
 - (int)textEffectsVisibilityLevel;
-- (void)updateFrame:(struct CGRect)arg1 withCorners:(unsigned long long)arg2;
+- (void)updateFrame:(struct CGRect)arg1 withCorners:(int)arg2;
 - (void)layoutSubviews;
 - (id)_generateCornerContentsImage:(int)arg1;
 - (void)transitionToStyle:(long long)arg1;
